make it better

+++ commands/say.js
@@ -3,6 +3,9 @@ exports.run = (client, message, args) => {
     return message.reply("sorry, you don't have enough permissions to use this!");
     // makes the bot say something and delete the message. As an example, it's open to anyone to use. 
     // To get the "message" itself we join the `args` back into a string with spaces: 
+    if (!args.length) {
+        return;
+    }
     message.delete();
     let botmessage = args.join(" ");
     message.channel.send(botmessage);
